+++ storybook/src/example-pages/Patronen/zoeken-in-lijst-cards.ts
@@ -34,12 +34,9 @@ examplePageFactory(
         <div>
           <h1>Zoeken in wetgeving</h1>
           <p>
-            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et
-            dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex
-            ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat
-            nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit
-            anim id est
-            laborum.${footnoteTemplate({
+            Gebruik deze pagina om te zoeken naar specifieke wetteksten, beleidsinformatie, of andersoortige regelgeving
+            die van invloed is op uw leefomgeving.
+            ${footnoteTemplate({
               number: 14,
               label: "Foreest Groen Consult (2018) Quickscan natuuronderzoek Ontwikkeling Herveld-Noord. 26 juli 2018.",
             })}