+++ storybook/src/example-pages/toepassingen/aanvragen/landingspagina.ts
@@ -37,8 +37,11 @@ examplePageFactory(
         ],
       })}
       <main>
-        <div class="row dso-banner dso-banner-implementation-specific-image" style="background-image: url('images/hands-on-trackpad.jpg')">
-          <div class="col-lg-6 col-sm-8">
+        <div
+          class="row dso-banner dso-banner-implementation-specific-image"
+          style="background-image: url('images/hands-on-trackpad.jpg')"
+        >
+          <div class="col-xs-12 col-lg-6 col-sm-8">
             ${highlightBoxTemplate({
               white: true,
               content: richContentTemplate({
@@ -56,7 +59,7 @@ examplePageFactory(
           </div>
         </div>
         <div class="row">
-          <div class="col-lg-8">
+          <div class="col-xs-12 col-lg-8">
             ${highlightBoxTemplate({
               yellow: true,
               content: richContentTemplate({
@@ -101,7 +104,7 @@ examplePageFactory(
               `,
             })}
           </div>
-          <div class="col-lg-4">
+          <div class="col-xs-12 col-lg-4">
             ${highlightBoxTemplate({
               border: true,
               white: true,
@@ -120,9 +123,9 @@ examplePageFactory(
         </div>
         <div class="row dso-featured">
           <div class="col-xs-12">
-            <h2>De stappen</h1>
+            <h2>De stappen</h2>
             <div class="row dso-equal-heights">
-              <div class="col-md-3 col-sm-6">
+              <div class="col-xs-12 col-md-3 col-sm-6">
                 ${highlightBoxTemplate({
                   white: true,
                   dropShadow: true,
@@ -138,7 +141,7 @@ examplePageFactory(
                   }),
                 })}
               </div>
-              <div class="col-md-3 col-sm-6">
+              <div class="col-xs-12 col-md-3 col-sm-6">
                 ${highlightBoxTemplate({
                   white: true,
                   dropShadow: true,
@@ -151,7 +154,7 @@ examplePageFactory(
                   }),
                 })}
               </div>
-              <div class="col-md-3 col-sm-6">
+              <div class="col-xs-12 col-md-3 col-sm-6">
                 ${highlightBoxTemplate({
                   white: true,
                   dropShadow: true,
@@ -164,7 +167,7 @@ examplePageFactory(
                   }),
                 })}
               </div>
-              <div class="col-md-3 col-sm-6">
+              <div class="col-xs-12 col-md-3 col-sm-6">
                 ${highlightBoxTemplate({
                   white: true,
                   dropShadow: true,
@@ -182,7 +185,7 @@ examplePageFactory(
               </div>
             </div>
             <div class="row dso-equal-heights">
-              <div class="col-md-3 col-sm-6">
+              <div class="col-xs-12 col-md-3 col-sm-6">
                 ${highlightBoxTemplate({
                   white: true,
                   dropShadow: true,
@@ -195,7 +198,7 @@ examplePageFactory(
                   }),
                 })}
               </div>
-              <div class="col-md-3 col-sm-6">
+              <div class="col-xs-12 col-md-3 col-sm-6">
                 ${highlightBoxTemplate({
                   white: true,
                   dropShadow: true,
@@ -208,7 +211,7 @@ examplePageFactory(
                   }),
                 })}
               </div>
-              <div class="col-md-3 col-sm-6">
+              <div class="col-xs-12 col-md-3 col-sm-6">
                 ${highlightBoxTemplate({
                   white: true,
                   dropShadow: true,