+++ storybook/src/example-pages/toepassingen/samenwerken/samenwerken-overzicht.content.ts
@@ -13,6 +13,8 @@ export const header: Header = {
   authStatus: "loggedIn",
   loginUrl: "#login",
   logoutUrl: "#logout",
+  showHelp: true,
+  helpUrl: "#help",
   userProfileName: "J.A. Jansen",
   userProfileUrl: "#profiel",
 };